uncomment initMPPT() and clean up code

+++ include/IOManagement.h
@@ -48,14 +48,14 @@ void initData();
 void resetPID();
 
 // Resets duty cycle PID for specified array
-void resetArrayPID(int array);
+void resetArrayPID(uint8_t array);
 
 // Sets voltage output for all arrays
 // Value will be capped if outside V_MIN or V_MAX specified in const.h
-void setVoltOut(double voltage);
+void setVoltOut(float voltage);
 
 // Sets voltage output for specified array
-void setArrayVoltOut(double voltage, int array);
+void setArrayVoltOut(float voltage, uint8_t array);
 
 // Sets clearing of OV fault
 void clearOVFaultReset(uint8_t value);
